requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
paho-mqtt==1.6.1
pydantic==2.5.0
python-multipart==0.0.6
websockets==12.0
asyncio-mqtt==0.16.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1
redis==5.0.1
prometheus-client==0.19.0
structlog==23.2.0
python-dotenv==1.0.0
httpx==0.25.2

